Address nkcr's comment
